refactor: separate editor-specific code (#163)
