+++ src/newsletter-editor/editor/style.scss
@@ -1,0 +1,7 @@
+.editor-post-title {
+	display: none;
+}
+
+.components-button.editor-post-publish-button__button.is-primary {
+	display: none;
+}
