+++ src/newsletter-editor/layout/style.scss
