+++ src/newsletter-editor/sidebar/style.scss
@@ -1,10 +1,5 @@
 @import '~@wordpress/base-styles/colors';
 
-.wp-block-buttons .block-editor-block-list__layout {
-	display: flex;
-	justify-content: space-between;
-}
-
 .newspack-newsletters {
 	&__editor-block {
 		&__warning {
