+++ src/newsletter-editor/utils.js
