+++ webpack.config.js
@@ -10,15 +10,11 @@ const getBaseWebpackConfig = require( '@automattic/calypso-build/webpack.config.
 const path = require( 'path' );
 
 /**
- * Internal dependencies
- */
-// const { workerCount } = require( './webpack.common' ); // todo: shard...
-
-/**
  * Internal variables
  */
 const editor = path.join( __dirname, 'src', 'editor' );
 const admin = path.join( __dirname, 'src', 'admin' );
+const newsletterEditor = path.join( __dirname, 'src', 'newsletter-editor' );
 
 const webpackConfig = getBaseWebpackConfig(
 	{ WP: true },
@@ -26,6 +22,7 @@ const webpackConfig = getBaseWebpackConfig(
 		entry: {
 			editor,
 			admin,
+			newsletterEditor,
 		},
 		'output-path': path.join( __dirname, 'dist' ),
 	}
